+++ src/views/portal/index.js
@@ -4,7 +4,7 @@ import Layout from '../../components/layout/main/index'
 import SignIn from '../auth/signin/index'
 import SignUp from '../auth/signup/index'
 import Home from '../portal/home/index'
-import Top05 from '../portal/top05/index'
+import Top05 from '../portal/cards_list/index'
 
 import Error403 from '../error/403/index'
 import Error404 from '../error/404/index'
